TRAC-539: Finshed cleaning the code and writing comments
